==> 00_extraction_de_la_base_verification.ipynb <==
import pandas as pd
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)
import sqlite3

cnx = sqlite3.connect('_sources/harassement.db')
df = pd.read_sql_query("SELECT * FROM tweets1", cnx)

df.head(2)

len(df)

# not_strings = lambda x: True if (type(x) != str) else False
is_strings = lambda x: True if (type(x) == str) else False

df_clean = df[df.text.apply(is_strings)]

len(df_clean)

df = df.drop(['index', 'user_id'], axis=1)

df.head(2)

df_clean.rename(columns={'usernameTweet':'username_tweet', 'ID':'id'}, inplace=True)

df.head(2)

df_clean.to_pickle('_outputs/harassement.pickle')
