==> 01_clean_text.ipynb <==
import pandas as pd
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

from ftfy import fix_text
import html
import string
import unidecode
import re
import sys

import warnings
warnings.filterwarnings('ignore')

from ekphrasis.classes.segmenter import Segmenter
seg_tw = Segmenter(corpus="twitter")

import swifter

df = pd.read_pickle('_outputs/harassement.pickle')

df.head(2)

len(df)

remove_encoding_errors = lambda tweet: re.sub(r"\\\S+", r'', tweet, flags=re.MULTILINE)
# https://regex101.com/r/aIYcKR/1/

no_links_http = lambda tweet: re.sub(r'https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)', '', tweet, flags=re.MULTILINE) # no links
# no_www = lambda tweet: re.sub(r"www\.\S+", '', tweet, flags=re.MULTILINE)
no_www = lambda tweet: re.sub(r"www\S*", '', tweet, flags=re.MULTILINE) # cas où ça coupe après www

# https://regex101.com/r/M0MrfJ/2
no_http = lambda tweet: re.sub(r"http\S*", '', tweet, flags=re.MULTILINE) # cas où ça coupe après http

no_url_queries = lambda tweet: re.sub(r"([\S]*[?/&=%]\S+)", '', tweet, flags=re.MULTILINE)
# https://regex101.com/r/ORoXn5/5

no_remaining_http = lambda tweet: re.sub(r"https?", '', tweet, flags=re.MULTILINE)
# https://regex101.com/r/Cr9IzK/2/

punct_as_separators = lambda tweet: re.sub(r"(\S)[#,.+*](\S)", r'\1 \2', tweet, flags=re.MULTILINE)
# https://regex101.com/r/9xjwh0/1/

all_hasgtags = lambda tweet: re.findall(r"#(\w+)", tweet)
all_mentions_hashtags_with_spaces = lambda tweet: re.findall(r"[@#]\s+\S+", tweet, flags=re.MULTILINE)
# https://regex101.com/r/wB8CSx/2/

no_mentions = lambda tweet: re.sub(r"@(\w+)", ' ', tweet, flags=re.MULTILINE)

no_digits = str.maketrans(string.digits, ' '*len(string.digits))

# https://stackoverflow.com/questions/34860982/replace-the-punctuation-with-whitespace
punctuation_to_space = lambda tweet: re.sub(r"[(){}\"'’,.;@#?!&%$/\\]+\ *", " ", tweet)

no_more_than_20 = lambda tweet: ' '.join(word for word in tweet.split() if len(word) <= 20)
no_less_than_3 = lambda tweet: ' '.join(word for word in tweet.split() if len(word) >= 3)

no_files_extensions_and_too = lambda tweet : re.sub(r'\bhtml\b|\bphp\b|\btoo\b', " ", tweet)

def clean_hashtags_mentions(tweet):
    mh_old = all_mentions_hashtags_with_spaces(tweet)
    mh_new = [m.replace(' ', '') for m in mh_old]
    tuples = zip(mh_old, mh_new)
    for t in tuples:
        tweet =tweet.replace(t[0], t[1])
    return tweet

def convert_hashtags(tweet):
    hasgtags = all_hasgtags(tweet)
    for h in hasgtags:
        h_extended = seg_tw.segment(h)
        tweet = tweet.replace('#{}'.format(h),h_extended)
    return tweet

# https://stackoverflow.com/a/49146722/330558
def remove_emoji(string):
    emoji_pattern = re.compile("["
                           u"\U0001F600-\U0001F64F"  # emoticons
                           u"\U0001F300-\U0001F5FF"  # symbols & pictographs
                           u"\U0001F680-\U0001F6FF"  # transport & map symbols
                           u"\U0001F1E0-\U0001F1FF"  # flags (iOS)
                           u"\U00002702-\U000027B0"
                           u"\U000024C2-\U0001F251"
                           "]+", flags=re.UNICODE)
    return emoji_pattern.sub(r'', string)

def clean_text(tweet):
    
    # Formatage
    tweet = remove_encoding_errors(tweet) # \\xe2\\x80\\xb
    tweet = fix_text(tweet) # Fix unicode
    ## fix_text('uÌˆnicode') -> ünicode
    tweet = html.unescape(tweet) # Fix unescape html special caracters
    
    tweet = remove_emoji(tweet)
    
    tweet = tweet.replace('\n', ' ').replace('\r', '') # replace line breaks with spaces, removes carriage returns

    tweet = tweet.strip() # Strip spaces
    
    tweet = tweet.lower()
        
    # Liens OK
    tweet = no_links_http(tweet) # No links with http(s)
    tweet = no_www(tweet) # no remaining www
    tweet = no_http(tweet) # no remaining http
    
    # Liens cassés
    tweet = no_url_queries(tweet) # No url queries when there ara spaces between the http(s) and url queries
    tweet = no_remaining_http(tweet) # remaining http(s)
    
    #listes : win,food,clothes,groceries
    tweet = punct_as_separators(tweet)
    
    #hashtags
    tweet = clean_hashtags_mentions(tweet) # when there are spaces between # or @ and the text
    tweet = convert_hashtags(tweet) # expand hashtags
    
    # remove mentions
    tweet = no_mentions(tweet)
    
    # no digits
    tweet = tweet.translate(no_digits)
    
    tweet = punctuation_to_space(tweet)
    
    tweet = no_more_than_20(tweet) # no more than 20 characters
    
    tweet = no_less_than_3(tweet)
    
    tweet = no_files_extensions_and_too(tweet) # remove file extensions and too of #metoo since hashtags are expanded
    
    tweet = " ".join(tweet.split()) #remove multiple spaces
    
    return tweet

df['clean_text'] = df.text.swifter.apply(clean_text)

df.head(2)

df[:20]['clean_text']

df.to_pickle('_outputs/harassement_clean.pickle')
