==> 02_detect_french_and_more_than_20_characters.ipynb <==
import pandas as pd
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

import warnings
warnings.filterwarnings('ignore')

from langdetect import detect
import swifter

df = pd.read_pickle('_outputs/harassement_clean.pickle')

df.head(2)

len(df)

sup_20 = lambda x: len(x) > 20
inf_20 = lambda x: len(x) < 20

df[df.clean_text.apply(inf_20)].head(2)

df = df[df.clean_text.apply(sup_20)]

len(df)

df['lang'] = df.clean_text.swifter.apply(lambda x: detect(x))

df[df.lang != 'fr'].head(2)

list(df[df.id == '1044325242766807042']['clean_text'])

df = df[df.lang == 'fr']

len(df)

df.to_pickle('_outputs/harassement_clean_french.pickle')
